rename bundle utils
