+++ src/index.ts
@@ -3,6 +3,6 @@ import * as IfhirR4 from './FHIR-R4/interfaces/IModel';
 import * as fhirR3 from './FHIR-R3';
 import * as dstu2 from './FHIR-DSTU2';
 import { ResourceUtilities } from './library/ResourceUtils/ResourceUtilities';
-import { BundleUtilities } from './library/BundleUtils/BundleUtilities';
+import { BundleUtilities } from './library/BundleUtils/BundleUtils';
 export { fhirR4, fhirR3, IfhirR4, dstu2, ResourceUtilities, BundleUtilities };
